+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 ### :house: Internal
 - `toolbars`
   - [#296](https://github.com/wix-incubator/rich-content/pull/296) improve performance by eliminating a lot of unnecessary work that was done on the toolbars for every keystroke.
+- `examples`
+  - [#298](https://github.com/wix-incubator/rich-content/pull/296) hmr + error handling for viewer
 <hr/>
 
 ## 3.3.3 (Apr 30, 2019)